+++ contract-testing/CHANGELOG.md
@@ -3,12 +3,13 @@
 ## Unreleased changes
 
 - Add methods to `Chain` for setting the exchange rates and block time based on queries from an external node.
-  - Add the method `setup_external_node_connection` for configuring the external node to use.
+  - Add a builder pattern via `Chain::builder` for setting up the connection to an external node and other parameters.
   - Add the method `set_exchange_rates_via_external_node` for setting the exchange rates.
   - Add the method `set_block_time_via_external_node` for setting the block time.
-  - Add the method `block_time` to get the current block time.
+  - Add the methods `set_block_time` and `block_time` to set and get the block time.
   - Add the method `sticky_block` for viewing the "sticky block", a concept explained in the `*via_external_node` methods.
 - Add the `BlockHash` type (re-exported from internal crate) for use in the methods above.
+- Expose the `Timestamp` type, which `SlotTime` is an alias of.
 
 ## 2.0.0